%%  WARNING: This file was automatically generated from the associated 
%%  sigseries.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the sigseries.mid file in the project MID directory. Once the 
%%  sigseries.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the asSeriesData function

\name{asSeriesData}
\alias{asSeriesData}
\title{Convert to a seriesData}
\concept{splus series}
\usage{asSeriesData(x)}
\description{Essentially just converts data to rectangular form, if possible.}

\arguments{
\item{x}{input object to coerce.}

}

\value{
rectangular conversion of input argument.
}
\seealso{
\code{\link{numericSequence}}, \code{\link{seriesData}}, \code{\link{seriesDataNew}}, \code{\link{seriesDataValid}}, \code{\link{signalSeries}}, \code{\link{as.data.frame.signalSeries}}, \code{\link{as.matrix.signalSeries}}, \code{\link{cumsum.signalSeries}}, \code{\link{deltat.signalSeries}}, \code{\link{diff.signalSeries}}, \code{\link{plot.signalSeries}}.}

\keyword{ts}

