\name{bits.per.integer}
\alias{bits.per.integer}
\title{ Internal Size of an integer }
\description{
  Reports the size of an integer in bits.
}
\usage{
bits.per.integer()
}
\details{
  Included for S-PLUS compatibility; R currently uses 32-bit integers,
  even on 64-bit machines.
}
\author{ Dunlap, W.M., and Taylor, C.F. }
\examples{
bits.per.integer()
}
\keyword{ utilities }
