\name{numericSequence-methods}
\alias{numericSequence-methods}
\docType{methods}
\alias{Math,numericSequence-method}
\alias{Math2,numericSequence-method}
\alias{Ops,ANY,numericSequence-method}
\alias{Ops,numericSequence,ANY-method}
\alias{Summary,numericSequence-method}
\alias{[,numericSequence-method}
\alias{[<-,numericSequence-method}
\alias{[[,numericSequence-method}
\alias{[[<-,numericSequence-method}
\alias{coerce,numeric,numericSequence-method}
\alias{coerce,numericSequence,character-method}
\alias{coerce,numericSequence,integer-method}
\alias{coerce,numericSequence,numeric-method}
\alias{duplicated,numericSequence-method}
\alias{is.na,numericSequence-method}
\alias{length,numericSequence-method}
\alias{match,ANY,numericSequence-method}
\alias{match,numericSequence,ANY-method}
\alias{mean,numericSequence-method}
\alias{median,numericSequence-method}
\alias{quantile,numericSequence-method}
\alias{rev,numericSequence-method}
\alias{show,numericSequence-method}
\alias{sort,numericSequence,missing-method}
\alias{sort,numericSequence-method}
\alias{summary,numericSequence-method}
\alias{unique,numericSequence,missing-method}
\alias{unique,numericSequence-method}
\alias{which.na,numericSequence-method}
\title{Methods for class signalSeries in package splus2R}
\concept{numericSequence methods}
\description{Methods for class \code{signalSeries} in package \code{splus2R}}

\details{
%
Supported (generic) methods include:
\describe{
\item{as}{Target classes: character, integer, numeric.}

\item{Math}{Math functions}

\item{Math2}{Math2 functions}

\item{Ops}{Ops functions}

\item{Data access and replacement}{Single- and double-bracket access and replacement functions}

\item{duplicated}{Find duplciated entries in sequence}

\item{is.na}{Test for NA entries in sequence}

\item{length}{Length of sequence}

\item{match}{Matching functions for sequence}

\item{mean}{Mean of sequence}

\item{median}{Median of sequence}

\item{quantile}{Quantile of sequence}

\item{rev}{Reverse sequence}

\item{show}{Show the sequence}

\item{sort}{Sort the sequence}

\item{unique}{Find unique sequence entries}

\item{summary}{Summarize the sequence}

\item{unique}{Find unique elements of a sequence}

\item{which.na}{Find NA entries in the sequence}}
}
\seealso{
\code{\link{signalSeries}}.}

\keyword{methods}

