%%  WARNING: This file was automatically generated from the associated 
%%  sigseries.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the sigseries.mid file in the project MID directory. Once the 
%%  sigseries.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the seriesDataValid function

\name{seriesDataValid}
\alias{seriesDataValid}
\title{Validates the structure of seriesData object}
\concept{splus series}
\usage{seriesDataValid(object)}
\description{Checks to see if input object is rectangular.}

\arguments{
\item{object}{the object to check.}

}

\value{
logical value, TRUE is a valid seriesData object.
}
\seealso{
\code{\link{numericSequence}}, \code{\link{seriesData}}, \code{\link{asSeriesData}}, \code{\link{seriesDataNew}}, \code{\link{signalSeries}}, \code{\link{as.data.frame.signalSeries}}, \code{\link{as.matrix.signalSeries}}, \code{\link{cumsum.signalSeries}}, \code{\link{deltat.signalSeries}}, \code{\link{diff.signalSeries}}, \code{\link{plot.signalSeries}}.}

\keyword{ts}

