\name{signalSeries-methods}
\alias{signalSeries-methods}
\docType{methods}
\alias{Arith,signalSeries,ANY-method}
\alias{Compare,signalSeries,ANY-method}
\alias{Logic,signalSeries,ANY-method}
\alias{Math,signalSeries-method}
\alias{Math2,signalSeries-method}
\alias{Ops,signalSeries,ANY-method}
\alias{Summary,signalSeries-method}
\alias{[,signalSeries-method}
\alias{as.numeric,signalSeries-method}
\alias{as.vector,signalSeries-method}
\alias{coerce,list,signalSeries-method}
\alias{coerce,signalSeries,data.frame-method}
\alias{coerce,signalSeries,character-method}
\alias{coerce,signalSeries,complex-method}
\alias{coerce,signalSeries,integer-method}
\alias{coerce,signalSeries,logical-method}
\alias{coerce,signalSeries,matrix-method}
\alias{coerce,signalSeries,numeric-method}
\alias{coerce,signalSeries,vector-method}
\alias{dim,signalSeries-method}
\alias{length,signalSeries-method}
\alias{mean,signalSeries-method}
\alias{min,signalSeries-method}
\alias{ncol,signalSeries-method}
\alias{plot,signalSeries-method}
\alias{show,signalSeries-method}
\alias{sum,signalSeries-method}
\alias{summary,signalSeries-method}
\title{Methods for class signalSeries in package splus2R}
\concept{signalSeries methods}
\description{Methods for class \code{signalSeries} in package \code{splus2R}}

\details{
%
Supported (generic) methods include:
\describe{
\item{as}{Target classes: complex, character, matrix, numeric, logical, integer, vector, and data.frame.
S3 style methods include as.numeric and as.vector.}

\item{Arith}{Arithmetic functions}

\item{Compare}{Comparison functions}

\item{Logic}{Logical functions}

\item{Math}{Math functions}

\item{Math2}{Math2 functions}

\item{Ops}{Ops functions}

\item{dim}{Dimension of series (NULL is returned)}

\item{length}{Length of series}

\item{mean}{Mean of series}

\item{min}{Minimum of series}

\item{ncol}{Number of columns of series (1 is returned)}

\item{plot}{Plot the series}

\item{show}{Display the series}

\item{sum}{Sum the series.}

\item{summary}{Sumamrize the series object}}
}
\seealso{
\code{\link{numericSequence}}.}

\keyword{methods}

