\name{asSeriesData}
\alias{asSeriesData}
\title{Convert to a seriesData}
\concept{splus series}
\usage{asSeriesData(x)}
\description{Essentially just converts data to rectangular form, if possible.}

\arguments{
\item{x}{input object to coerce.}

}

\value{
rectangular conversion of input argument.
}
\seealso{
\code{\link{numericSequence}}, \code{\link{seriesData}}, \code{\link{seriesDataNew}}, \code{\link{seriesDataValid}}, \code{\link{signalSeries}}, \code{\link{as.data.frame.signalSeries}}, \code{\link{as.matrix.signalSeries}}, \code{\link{cumsum.signalSeries}}, \code{\link{deltat.signalSeries}}, \code{\link{diff.signalSeries}}, \code{\link{plot.signalSeries}}.}

\keyword{ts}

