\name{seriesDataNew}
\alias{seriesDataNew}
\title{Creates template for new seriesData object}
\concept{splus series}
\usage{seriesDataNew()}
\description{Creates template for new seriesData object., basically a NULL matrix.}


\value{
nULL matrix.
}
\seealso{
\code{\link{numericSequence}}, \code{\link{seriesData}}, \code{\link{asSeriesData}}, \code{\link{seriesDataValid}}, \code{\link{signalSeries}}, \code{\link{as.data.frame.signalSeries}}, \code{\link{as.matrix.signalSeries}}, \code{\link{cumsum.signalSeries}}, \code{\link{deltat.signalSeries}}, \code{\link{diff.signalSeries}}, \code{\link{plot.signalSeries}}.}

\keyword{ts}

