\name{seriesDataValid}
\alias{seriesDataValid}
\title{Validates the structure of seriesData object}
\concept{splus series}
\usage{seriesDataValid(object)}
\description{Checks to see if input object is rectangular.}

\arguments{
\item{object}{the object to check.}

}

\value{
logical value, TRUE is a valid seriesData object.
}
\seealso{
\code{\link{numericSequence}}, \code{\link{seriesData}}, \code{\link{asSeriesData}}, \code{\link{seriesDataNew}}, \code{\link{signalSeries}}, \code{\link{as.data.frame.signalSeries}}, \code{\link{as.matrix.signalSeries}}, \code{\link{cumsum.signalSeries}}, \code{\link{deltat.signalSeries}}, \code{\link{diff.signalSeries}}, \code{\link{plot.signalSeries}}.}

\keyword{ts}

